refactor: relocate SearchByCountryScreen.kt and its components
