+++ ui/src/main/java/com/example/ui/screens/search/searchByCountry/SearchByCountryScreenRoute.kt
@@ -1,4 +1,4 @@
-package com.example.ui.screens.searchByCountry
+package com.example.ui.screens.search.searchByCountry
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
