+++ viewModel/src/main/java/com/example/viewmodel/search/countrySearch/SearchByCountryInteractionListener.kt
@@ -3,5 +3,5 @@ package com.example.viewmodel.search.countrySearch
 interface SearchByCountryInteractionListener {
     fun onKeywordValueChanged(keyword: String)
     fun onCountrySelected(country: CountryUiState)
-    fun onRetryQuestClicked()
+    fun onRetryRequestClicked()
 }
