+++ viewModel/src/test/java/com/example/viewmodel/utils/entityHelper/createMovie.kt
@@ -1,6 +1,6 @@
 package com.example.viewmodel.utils.entityHelper
-import com.example.entity.Category
 import com.example.entity.Movie
+import com.example.entity.category.MovieGenre
 
 fun createMovie(
     id: Long = 1L,
@@ -8,7 +8,7 @@ fun createMovie(
     description: String = "A sample movie description.",
     poster: String = "https://example.com/poster.jpg",
     productionYear: Int = 2024,
-    categories: List<Category> = listOf(),
+    genres: List<MovieGenre> = listOf(),
     rating: Float = 4.5f
 ): Movie {
     return Movie(
@@ -17,7 +17,7 @@ fun createMovie(
         description = description,
         poster = poster,
         productionYear = productionYear,
-        categories = categories,
+        categories = genres,
         rating = rating,
         popularity = 0.0
     )
